complete part 1 task 3
To implement waitUntil(long x) method we use 2 variables (waketime) to
track the time and (callerThread) to track which thread to sleep. We
also use a binary semaphore to handle multiple call from other threads.

First the thread locks the semaphore and set the variables, then goes to
sleep. The timer interrupt checks the time every 500 ticks apart. If
waiting time for the thread is up then it wakes the caller thread and
reset the variables, also unlock the semaphore.