finish task 2 (implement condition variables directly)
Declare a queue of existing class FifoQueue to keep track of the waiting
threads. Also use e variable to keep track of the size of the queue as
the FifoQueue has no public method to get size.

Sleep:
Release the condition lock. Disable the machine interrupt. Add the current
thread to the queue and icrement size. Call the sleep method of the current
thread. Restore the machine interrupt status and the acquire the
condition lock.

Wake:
Disable the machine interrupt. Wake up a sleeping thread calling ready
method of the thread and decrement the size. Then restore the machine
interrupt.

WakeAll:
Call the wake method until the size of the waitQueue is zero.