complete part 2 task 2 multiprogramming
To allow multiprogramming use a LinkedList to track the free physical
pages in the UserKernel. Use a lock to ensure mutual exclution of the
free physical page access. To allow multiple processes to access the
memory, use a pageTable that maps the virtual page number to a physical
page number. When the process try to access the memory use the method
virtualToPhysicalAddress() to map the virtual address to appropiate
physical address. When loading the sections for the prosses allocate the
free pages for it so that no overlapping happens.